refactor(storage): extract transfer responsibility owner
